add deleting bought tickets
