+++ src/context/RoutesContext.js
@@ -1,6 +1,6 @@
 import React, {createContext, useState, useCallback, useEffect} from 'react';
 import apiUtils from '../utils/apiUtils';
-import {connectToRabbitMQ} from "../utils/rabbitMqUtils";
+
 
 export const RoutesContext = createContext();
 
@@ -8,22 +8,6 @@ export const RoutesProvider = ({ children }) => {
     const [routes, setRoutes] = useState([]);
     const [loading, setLoading] = useState(true);
     const [isFetching, setIsFetching] = useState(false);
-
-    useEffect(() => {
-        const topicName = 'ticket_exchange';
-        const queueName = 'ticket_response';
-
-        connectToRabbitMQ(topicName, queueName, handleRabbitMQMessage);
-    }, []);
-
-    const handleRabbitMQMessage = (message) => {
-        const { data } = JSON.parse(message.body);
-        console.log(message);
-        if (!isFetching) {
-            fetchRoutes();
-        }
-        setLoading(false);
-    };
 
     const changeLoading = (newLoading) => {
         setLoading(newLoading);
